feat(pages): add Cache page, to build/verify cache in browser
Moved cache script to a cache service, to be used by the blockchain store and the cache page, the
cache page receive the options to run the cache build script in the dapp and verifies the hash of
the config received with the target hash. It also provides the .json files of the configs and cache
to be downloaded.